requirements.txt
django>=5.0
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
python-dotenv
